Minor fixes and additions
i went ahead and removed a bunch of machines that did not to be where they were and put a ACTUAL stewards office in the ship, the vault door now can be locked with the lords key.  i also fixed some minor design details on the tower in he azure forest, making the tower look less stupid.