Fixes the sleeping icon, adds a description (#2436)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

As the title says, there was an unused sleeping icon since the
icon_state written was incorrect, this fixes it and adds a description
to the status.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Bug fix! And a description on top
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof of Testing (Required)


![image](https://github.com/user-attachments/assets/79d7db19-0299-4e6e-9169-c80c9e136ee9)

![image](https://github.com/user-attachments/assets/18950a9b-78b6-406d-befc-5b12b40cfa87)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->